Consolidated AI routes by removing duplicate router in openai/routes Updated main AI router to use OpenAIService with improved error handling Added OpenAI-specific error responses (rate limits, service unavailability) Simplified health check endpoint and removed unused routes Installed zod dependency for type validation
