+++ .gc/action-plans/07.09.25-1540-ai-route-cleanup-and-type-harmonization.md
@@ -31,12 +31,12 @@ Prepare the codebase for demo-day by simplifying the AI integration:
 | # | Status | Description | Files to modify |
 |---|--------|-------------|----------------|
 |1| ✓ |**Make `gameId` optional** in the Zod schema.|`backend/openai/types/game.ts`, `backend/src/types/index.ts`, any tests referencing `AnalyzeRequestSchema`.|
-|2| |**Role rename in backend**: change `'outsider_hint'→'encoder'`, `'insider_guess'→'decoder'`, `'encryptor'→'encoder'`, `'decryptor'→'decoder'`. |Key files: `backend/openai/types/game.ts`, `backend/src/types/index.ts`, `backend/src/game/*`, socket handlers.|
+|2| ✓ |**Role rename in backend**: change `'outsider_hint'→'encoder'`, `'insider_guess'→'decoder'`, `'encryptor'→'encoder'`, `'decryptor'→'decoder'`. |Key files: `backend/openai/types/game.ts`, `backend/src/types/index.ts`, `backend/src/game/*`, socket handlers.|
 |3| |**Delete duplicate router** and exports.|Remove `backend/openai/routes/ai.ts`; delete its import in any `index.ts` barrel.|
 |4| |**Remove MockAIService**.|Delete `backend/src/ai/mock.ts`; delete instantiation in `backend/src/routes/ai.ts`; update route to import `openAIService` from `backend/openai/services/openai.ts`.|
 |5| |**Update main AI route** to call real service and adjust validation (optional `gameId`).|`backend/src/routes/ai.ts`|
-|6| |**Consolidate Turn/Zod schemas**: keep canonical definition (current file) and import it everywhere else. Remove secondary re-exports.|Ensure only `backend/openai/types/game.ts` holds the definitions; update import paths in backend.|
-|7| |**Update tests** for new role names & optional gameId.|All Jest files in `backend/tests`.|
+|6| ✓ |**Consolidate Turn/Zod schemas**: keep canonical definition (current file) and import it everywhere else. Remove secondary re-exports.|Ensure only `backend/openai/types/game.ts` holds the definitions; update import paths in backend.|
+|7| ✓ |**Update tests** for new role names & optional gameId.|All Jest files in `backend/tests`.|
 |8| |**Create opt-in integration test** (skipped by default).|Add `tests/aiIntegrationReal.test.ts` that is skipped unless `process.env.OPENAI_E2E==='true'`.|
 |9| |**Documentation cleanup**.|Update `/backend/ai-api-docs.md` & `/backend/server-spec.md` with the new schema; remove mention of MockAIService.|
